Refactor naming of internal render packages
