+++ render/internal/context/text_state.go
